refactor(span)!: export `ContentEq` trait from root of `oxc_span` crate (#8869)
Export `ContentEq` as `oxc_span::ContentEq`. Previously was within a sub-module `oxc_span::cmp::ContentEq`. This made sense when we also had `ContentHash`, but now that's removed, it makes more sense for `ContentEq` to be exported from root of the crate.
