+++ crates/oxc_ast_macros/src/ast.rs
@@ -80,7 +80,7 @@ fn abs_trait(
     } else if ident == "GetAddress" {
         (quote!(::oxc_allocator::GetAddress), TokenStream::default())
     } else if ident == "ContentEq" {
-        (quote!(::oxc_span::cmp::ContentEq), TokenStream::default())
+        (quote!(::oxc_span::ContentEq), TokenStream::default())
     } else if ident == "ESTree" {
         (quote!(::oxc_estree::ESTree), TokenStream::default())
     } else {
