+++ crates/oxc_linter/src/rules/eslint/no_duplicate_case.rs
@@ -1,7 +1,7 @@
 use oxc_ast::ast::Expression;
 use oxc_diagnostics::OxcDiagnostic;
 use oxc_macros::declare_oxc_lint;
-use oxc_span::{cmp::ContentEq, GetSpan, Span};
+use oxc_span::{ContentEq, GetSpan, Span};
 
 use crate::{context::LintContext, rule::Rule, AstNode};
