+++ crates/oxc_linter/src/rules/eslint/no_useless_call.rs
@@ -4,7 +4,7 @@ use oxc_ast::{
 };
 use oxc_diagnostics::OxcDiagnostic;
 use oxc_macros::declare_oxc_lint;
-use oxc_span::cmp::ContentEq;
+use oxc_span::ContentEq;
 use oxc_span::Span;
 
 use crate::{context::LintContext, rule::Rule, AstNode};
