+++ crates/oxc_linter/src/rules/eslint/prefer_spread.rs
@@ -4,7 +4,7 @@ use oxc_ast::{
 };
 use oxc_diagnostics::OxcDiagnostic;
 use oxc_macros::declare_oxc_lint;
-use oxc_span::{cmp::ContentEq, Span};
+use oxc_span::{ContentEq, Span};
 
 use crate::{ast_util::is_method_call, context::LintContext, rule::Rule, AstNode};
