+++ crates/oxc_minifier/src/peephole/minimize_conditional_expression.rs
@@ -1,5 +1,5 @@
 use oxc_ast::{ast::*, NONE};
-use oxc_span::{cmp::ContentEq, GetSpan};
+use oxc_span::{ContentEq, GetSpan};
 use oxc_syntax::es_target::ESTarget;
 
 use crate::ctx::Ctx;
