+++ crates/oxc_minifier/src/peephole/minimize_conditions.rs
@@ -1,6 +1,6 @@
 use oxc_ast::ast::*;
 use oxc_ecmascript::{constant_evaluation::ValueType, ToInt32};
-use oxc_span::{cmp::ContentEq, GetSpan};
+use oxc_span::{ContentEq, GetSpan};
 use oxc_syntax::es_target::ESTarget;
 
 use crate::ctx::Ctx;
